+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 
 - - -
 ## v0.4.0-dev (current)
-- (2026 01 31): Finish writing 9-slice UI
+
+#### changes
+- (2026 01 31): Add text window restriction
 - (2026 01 31): Fix missing `fsl_render.ndc_scale` at init time
+- (2026 01 31): Add 9-slice UI
 - (2026 01 31): Add test `nine_slice`
 - (2026 01 31): Split `fsl_str_libs[][]` into internal and external arrays
 - (2026 01 31): Add internal framebuffer size callback and an optional pointer
@@ -40,8 +43,6 @@
 - (2026 01 29): Change `CHANGELOG.md` format to include date in changes
 - (2026 01 29): Change date format in text files `(YYYY MMM DD)` ->
                 `(YYYY MM DD)`
-
-#### changes
 
 - - -
 ## v0.4.0-beta (2026 Jan 29)